+++ src/main/java/com/github/hokkaydo/eplbot/module/code/RustCompiler.java
@@ -4,9 +4,10 @@ import java.io.File;
 import java.io.FileWriter;
 import java.io.InputStream;
 
-public class RustCompiler {
+public class RustCompiler implements Runner{
     private static final String CURRENT_DIR = System.getProperty("user.dir") + "\\src\\temp\\";
-    public static String run(String input, Integer runTimeout) {
+    @Override
+    public String run(String input, Integer runTimeout) {
         if (containsUnsafeKeywords(input)){
             return "Compilation failed:\nCheck if 'std' or 'use' are used";
         }
@@ -36,9 +37,9 @@ public class RustCompiler {
                 Process run = new ProcessBuilder(new File(CURRENT_DIR, executableFileName).getAbsolutePath()).directory(new File(CURRENT_DIR)).redirectErrorStream(true).start();
                 Thread timeoutThread = new Thread(() -> {
                     try {
-                        Thread.sleep(1000*runTimeout);
+                        Thread.sleep(1000L*runTimeout);
                         run.destroy();
-                    } catch (InterruptedException e) {}
+                    } catch (InterruptedException ignored) {}
                 });
                 timeoutThread.start();
                 outputStream.reset();
@@ -63,7 +64,7 @@ public class RustCompiler {
                 }
             } else {
                 deleteFiles();
-                return "Compilation failed:\n" + outputStream.toString();
+                return "Compilation failed:\n" + outputStream;
             }
         } catch (Exception e) {
             e.printStackTrace();
@@ -80,18 +81,24 @@ public class RustCompiler {
         return false;
     }
     private static void deleteFiles(){
-        for (File file : new File(CURRENT_DIR).listFiles()) {
-            if (file.isFile() && file.getName().startsWith("temp")) {
-                String[] validExtensions = {".exe", ".pdb", ".rs"};
-                boolean shouldDelete = false;
-                for (String extension : validExtensions) {
-                    if (file.getName().equals("temp"+extension)) {
-                        shouldDelete = true;
-                        break;
+        File outputDirectory = new File(CURRENT_DIR);
+        File[] files = outputDirectory.listFiles();
+        if (files == null){
+            System.out.println("NPE trying to delete created files");
+        } else {
+            for (File file : new File(CURRENT_DIR).listFiles()) {
+                if (file.isFile() && file.getName().startsWith("temp")) {
+                    String[] validExtensions = {".exe", ".pdb", ".rs"};
+                    boolean shouldDelete = false;
+                    for (String extension : validExtensions) {
+                        if (file.getName().equals("temp"+extension)) {
+                            shouldDelete = true;
+                            break;
+                        }
                     }
-                }
-                if (shouldDelete) {
-                    file.delete();
+                    if (shouldDelete) {
+                        file.delete();
+                    }
                 }
             }
         }
